chore: change Run Configurations for Qodana
This commit modifies the Run Qodana.run.xml file. The script parameter was removed, and instead, the "cleanInspections" and "runInspections" options were added to the taskNames list of the Gradle external system. Additionally, the RunAsTest tag was added as false. There were no changes at the end of the file.