refactor: modify the project generator to include project validation
This commit modifies the `NuxtCliProjectGenerator` to include project validation. The `validateProjectPath` method now checks if the project name is a valid NPM package name using the `NodePackageUtil.validateNpmPackageName` method. If the name is not valid, an error message is returned. Otherwise, the validation continues as before. This change helps ensure that only valid project names are used when generating Nuxt.js projects.