refactor: update project generator for simplicity and clearness
Simplified the syntax of the filters, generatorArgs and generateInTemp methods in NuxtCliProjectGenerator.kt file by removing redundant declarations of returning types, which are implicit from the Kotlin language point of view. The validateProjectPath method was removed as the path validation will be performed at a higher level, optimizing the code execution flow.