chore: change plugin repository URL in gradle properties
The plugin repository URL has been updated in the 'gradle.properties' file. This change is due to the shifting of the IntelliJ Nuxt.js Plugin repository from the 'nekofar' account to the 'KartanHQ' account on GitHub. This update ensures the build script points to the correct and current location of the plugin repository.